Add support for custom per-view colors
In general, this implements an indirection from the
config->border_colors singleton widely used for decoration colors.
Containers may be a more sensible target than views, but this is enough
to demonstrate a proof of concept.

While this change is expected to be mostly aesthetic, there are some
practical uses of such features; e.g. Qubes uses decoration colors to
signal trust domains of running applications to the user, and similar
musings could apply to clear, visual alerts of user interactions (e.g.,
making the Tridactyl native editor window have a distinct appearance).