Add Security Labels as a criterion
Security labels (sometimes "contexts") are process attributes whose
meaning is defined by a loaded Linux Security Module, if any. Permitting
these to be specified as a criterion, in particular, allows
differentiating behavior of process based on their "security domain",
even if those processes are otherwise identical (e.g., the same binary,
the same window title, etc). This could prove useful, e.g., to ensure
"trusted" and "untrusted" applications appear visually distinct.

While this patch is standalone, it is based on (and highly recommended
to be used with) the window-colors patch, which permits a use case of
visual distinction by decoration color.